fix: change source of truth for types to umd/index.d.ts to resolve issue with typescript import in 5.x
